Always allocate input buffer on heap, prevent copying CsvParser
Previously, if constructed on the stack, the CsvParser object would
occupy over 128 KiB of stack, making it prone to cause stack overflows
when using multiple instances.

Also, if the CsvParser object was copied and both objects were continued
to be used this would result in some undefined behavior,
since both instances were using the same underlying istream.
By preventing copying the object (implicitly, through the unqiue_ptr member)
users are forced to either pass a reference or to make use of std::move